Update linear-inf paper

+++ build.sh
@@ -2,10 +2,10 @@
 
 path=$(nix build --json --no-link | jq -r .[0].outputs.out)
 
-if [ $? -ne 0 ]; then
+if [ "$path" == "" ]; then
     echo "Nix failed"
+    exit 1
 fi
-
 rm -r ./alexarice.github.io/*
 cp -L -r $path/* ./alexarice.github.io
 chmod u+rw -R ./alexarice.github.io/*
